requirements.txt
fastapi
requests
PyYAML
uvicorn
